+++ src/Animation.hpp
@@ -2,8 +2,8 @@
 #define __ANIMATION_HPP__
 
 typedef struct S_Animation {
-  unsigned int  frame;
-  sf::Time      time;
+  uint16_t    frame;
+  sf::Time    time;
 } Animation;
 
 #endif // __ANIMATION_HPP__
